cuda works

+++ src_cuda/math/vec3.cuh
@@ -3,9 +3,9 @@
 #include <cmath>
 
 struct Vec3 {
-    static constexpr const double eps = 0.000001;
+    static constexpr const float eps = 0.000001f;
 
-    double x, y, z;
+    float x, y, z;
 
     __host__ __device__ constexpr Vec3 operator-() const { return {-x, -y, -z}; }
 
@@ -31,22 +31,22 @@ struct Vec3 {
         return *this;
     }
 
-    __host__ __device__ constexpr Vec3 operator*(double f) const { return {x * f, y * f, z * f}; }
+    __host__ __device__ constexpr Vec3 operator*(float f) const { return {x * f, y * f, z * f}; }
 
-    __host__ __device__ constexpr friend Vec3 operator*(double f, const Vec3 v) {
+    __host__ __device__ constexpr friend Vec3 operator*(float f, const Vec3 v) {
         return {v.x * f, v.y * f, v.z * f};
     }
 
-    __host__ __device__ Vec3& operator*=(double f) {
+    __host__ __device__ Vec3& operator*=(float f) {
         x *= f;
         y *= f;
         z *= f;
         return *this;
     }
 
-    __host__ __device__ constexpr Vec3 operator/(double f) const { return {x / f, y / f, z / f}; }
+    __host__ __device__ constexpr Vec3 operator/(float f) const { return {x / f, y / f, z / f}; }
 
-    __host__ __device__ Vec3& operator/=(double f) {
+    __host__ __device__ Vec3& operator/=(float f) {
         x /= f;
         y /= f;
         z /= f;
@@ -72,15 +72,15 @@ struct Vec3 {
         return x != other.x || y != other.y || z != other.z;
     };
 
-    __host__ __device__ constexpr double lengthSquared() const { 
+    __host__ __device__ constexpr float lengthSquared() const { 
         return x * x + y * y + z * z; 
     }
     
-    __host__ __device__ double length() const { 
+    __host__ __device__ float length() const { 
         return sqrt(lengthSquared()); 
     }
 
-    __host__ __device__ constexpr double dot(const Vec3& other) const {
+    __host__ __device__ constexpr float dot(const Vec3& other) const {
         return x * other.x + y * other.y + z * other.z;
     }
 
@@ -91,7 +91,7 @@ struct Vec3 {
     }
 
     __host__ __device__ Vec3 unitVector() const {
-        double length = this->length();
+        float length = this->length();
         if (length > eps)
             return (*this) * (1 / this->length());
         else return {0,0,0};
